+++ src/components/collections/collection-modal.js
@@ -1,20 +1,22 @@
 import { PolymerElement, html } from "@polymer/polymer/polymer-element.js";
-//import "@polymer/polymer/lib/elements/dom-repeat.js";
+
 import "@vaadin/vaadin-grid/vaadin-grid";
 import "@vaadin/vaadin-grid/vaadin-grid-sorter";
 
-import { connect } from 'pwa-helpers/connect-mixin.js';
+import { connect } from "pwa-helpers/connect-mixin.js";
 
 // Load redux store
-import { store } from '../../store.js';
+import { store } from "../../store.js";
 
 // These are the actions needed by this element.
-import { UPDATE_COLLECTION_MODAL } from '../../actions/collections.js';
+import { UPDATE_COLLECTION_MODAL } from "../../actions/collections.js";
 
 // Lazy load reducers
-import collections from '../../reducers/collections.js';
+import collections from "../../reducers/collections.js";
+import settings from "../../reducers/settings.js";
 store.addReducers({
-  collections
+  collections,
+  settings
 });
 
 import "../utilities/gwn-modal.js";
@@ -60,11 +62,11 @@ class CollectionModal extends connect(store)(PolymerElement) {
         vaadin-cell-grid-content {
           height: 2.5rem;
         }
-        
+
         .icon {
           width: 32px;
           height: 32px;
-          margin-right: .5rem;
+          margin-right: 0.5rem;
           vertical-align: middle;
         }
 
@@ -78,17 +80,22 @@ class CollectionModal extends connect(store)(PolymerElement) {
       <gwn-modal hidden="[[!open]]" on-hidden-changed="_hiddenChanged">
         <h3 class="headline" slot="title">[[collectionName]]</h3>
         <div slot="content">
-          <vaadin-grid theme="no-border row-stripes" height-by-rows items="[[collectionItems]]">
+          <vaadin-grid
+            theme$="no-border row-stripes [[theme]]"
+            height-by-rows
+            items="[[collectionItems]]"
+          >
             <vaadin-grid-column>
               <template class="header">
                 <vaadin-grid-sorter path="name">Item</vaadin-grid-sorter>
               </template>
               <template>
-                <gwn-item-icon 
-                  class="icon" 
-                  name="[[item.name]]" 
-                  icon="[[item.icon]]" 
-                  rarity="[[item.rarity]]"></gwn-item-icon>
+                <gwn-item-icon
+                  class="icon"
+                  name="[[item.name]]"
+                  icon="[[item.icon]]"
+                  rarity="[[item.rarity]]"
+                ></gwn-item-icon>
                 <span>[[ item.name ]]</span>
               </template>
               <template class="footer">
@@ -99,28 +106,46 @@ class CollectionModal extends connect(store)(PolymerElement) {
             <vaadin-grid-column>
               <template class="header">
                 <div class="align-right">
-                  <vaadin-grid-sorter path="buys.unit_price">Buy Order</vaadin-grid-sorter>
+                  <vaadin-grid-sorter path="buys.unit_price"
+                    >Buy Order</vaadin-grid-sorter
+                  >
                 </div>
               </template>
               <template>
-                <gw2-coin-output prepend-zeroes coin-string="[[item.buys.unit_price]]"></gw2-coin-output>
+                <gw2-coin-output
+                  prepend-zeroes
+                  coin-string="[[item.buys.unit_price]]"
+                ></gw2-coin-output>
               </template>
               <template class="footer">
-                <gw2-coin-output class="text-bold" prepend-zeroes coin-string="[[totalBuy]]"></gw2-coin-output>
+                <gw2-coin-output
+                  class="text-bold"
+                  prepend-zeroes
+                  coin-string="[[totalBuy]]"
+                ></gw2-coin-output>
               </template>
             </vaadin-grid-column>
 
             <vaadin-grid-column>
               <template class="header">
                 <div class="align-right">
-                  <vaadin-grid-sorter path="sells.unit_price">Sell Listing</vaadin-grid-sorter>
+                  <vaadin-grid-sorter path="sells.unit_price"
+                    >Sell Listing</vaadin-grid-sorter
+                  >
                 </div>
               </template>
               <template>
-                <gw2-coin-output prepend-zeroes coin-string="[[item.sells.unit_price]]"></gw2-coin-output>
+                <gw2-coin-output
+                  prepend-zeroes
+                  coin-string="[[item.sells.unit_price]]"
+                ></gw2-coin-output>
               </template>
               <template class="footer">
-                <gw2-coin-output class="text-bold" prepend-zeroes coin-string="[[totalSell]]"></gw2-coin-output>
+                <gw2-coin-output
+                  class="text-bold"
+                  prepend-zeroes
+                  coin-string="[[totalSell]]"
+                ></gw2-coin-output>
               </template>
             </vaadin-grid-column>
           </vaadin-grid>
@@ -138,7 +163,8 @@ class CollectionModal extends connect(store)(PolymerElement) {
       collectionName: String,
       collectionItems: String,
       totalBuy: Number,
-      totalSell: Number
+      totalSell: Number,
+      theme: String
     };
   }
 
@@ -154,11 +180,12 @@ class CollectionModal extends connect(store)(PolymerElement) {
   // This is called every time something is updated in the store.
   _stateChanged(state) {
     if (!state) return;
-    this.set('open', state.collections.collectionModalOpened);
-    this.set('collectionName', state.collections.selectedCollection.name);
-    this.set('collectionItems', state.collections.selectedCollection.items);
-    this.set('totalBuy', state.collections.selectedCollection.totalBuy);
-    this.set('totalSell', state.collections.selectedCollection.totalSell);
+    this.set("open", state.collections.collectionModalOpened);
+    this.set("collectionName", state.collections.selectedCollection.name);
+    this.set("collectionItems", state.collections.selectedCollection.items);
+    this.set("totalBuy", state.collections.selectedCollection.totalBuy);
+    this.set("totalSell", state.collections.selectedCollection.totalSell);
+    this.set("theme", state.settings.theme);
   }
 }